Реализован EventViewController вместо HabitViewController и IrregularEventViewController
