+++ Tracker/EventViewCpntroller.swift
@@ -1,0 +1,8 @@
+//
+//  EventViewCpntroller.swift
+//  Tracker
+//
+//  Created by Aleksei Frolov on 16.11.2024.
+//
+
+import Foundation
