+++ Tracker/TrackerTypeSelectionViewController.swift
@@ -84,14 +84,14 @@ final class TrackerTypeSelectionViewController: UIViewController {
     
     @objc private func habitButtonTapped() {
         guard navigationController?.topViewController == self else { return }
-        let habitVC = HabitViewController()
+        let habitVC = EventViewController(trackerType: .habit)
         navigationController?.pushViewController(habitVC, animated: true)
         print("Переход на экран создания Привычки")
     }
     
     @objc private func irregularEventButtonTapped() {
         guard navigationController?.topViewController == self else { return }
-        let irregularEventVC = IrregularEventViewController()
+        let irregularEventVC = EventViewController(trackerType: .irregularEvent)
         navigationController?.pushViewController(irregularEventVC, animated: true)
         print("Переход на экран создания Нерегулярного события")
     }
